Created SQL View, Models, Controller, Angular Service and component for displaying weather grid

+++ Project/DBProject/Models/WeatherDetails/Main.cs
@@ -15,15 +15,15 @@ namespace DBProject.Models.WeatherDetails
         [Column("city_id")]
         public int CityId { get; set; }
         [Column("temp")]
-        public float Temp { get; set; }
+        public double Temp { get; set; }
         [Column("pressure")]
         public int Pressure { get; set; }
         [Column("humidity")]
         public int Humidity { get; set; }
         [Column("temp_min")]
-        public float Temp_min { get; set; }
+        public double Temp_min { get; set; }
         [Column("temp_max")]
-        public float Temp_max { get; set; }
+        public double Temp_max { get; set; }
         public int TempC { get => (int)Math.Round(Temp - 273.15); }
     }
 }
